refractor(rename_directories): rename directories to capitalize, and moved API url constant to fuse constants files
