+++ Constants/ConnectionConstants.js
@@ -7,3 +7,4 @@ export const O_CONNECTION_STATUS_CONSTANTS = {
 export const O_WEB_SOCKET_CONSTANTS = {
     "url": "wss://ws.hothothot.dog:9502",
 }
+export const S_API_URL = "https://hothothot.dog/api/capteurs/";
