+++ Entity/TemperatureEntity.js
