+++ Views/assets/script/script.js
