--- constants/UrlConstants.js
@@ -1,3 +1,0 @@
-export const S_API_URL = "https://hothothot.dog/api/capteurs/";
-
-export const S_WEBSOCKET_URL = "wss://ws.hothothot.dog:9502";
